organized code in components and functions
