+++ editor/src/app/form/index.js
@@ -42,7 +42,6 @@ const Liform = props => {
   );
 };
 
-
 export default Liform;
 
 export {
