+++ editor/src/app/utils/calls.js
@@ -1,0 +1,5 @@
+export const getReleases = (versionsUrl) => {
+  return fetch(versionsUrl)
+    .then(res => res.json())
+    .then(data => data.map(d => d.name));
+};
