+++ editor/src/asset/block.scss
@@ -8,6 +8,7 @@
   display: flex;
   flex-wrap: wrap;
   padding: 10px;
+  min-height: 250px;
 }
 
 .block_heading_oval {
